+++ frontend/src/pages/Landing/LandingPage.tsx
@@ -18,6 +18,7 @@ import type { Article } from '../News/NewsPage'
 import { ArticlesSkeleton } from '../News/AllPosts'
 import { usePrefersReducedMotion } from '../../utils/hooks/usePrefersReducedMotion'
 import { urlMap } from '../../utils/externalUrls'
+import HetBigCTA from '../../styles/HetComponents/HetBigCTA'
 
 function LandingPage() {
   const { isLoading, error, data }: any = useQuery(
@@ -74,14 +75,9 @@ function LandingPage() {
               equity and justice for all.
             </p>
             <div className='mb-10 mt-10 lg:mt-20'>
-              <Button
-                id='landingPageCTA'
-                variant='contained'
-                className='rounded-2xl px-8 py-5 text-exploreButton text-white'
-                href={EXPLORE_DATA_PAGE_LINK}
-              >
+              <HetBigCTA id='landingPageCTA' href={EXPLORE_DATA_PAGE_LINK}>
                 Explore the data
-              </Button>
+              </HetBigCTA>
             </div>
           </div>
           <div className='w-full border-0 border-l border-solid px-12 py-4 md:w-5/12'>
@@ -247,13 +243,9 @@ function LandingPage() {
           </ul>
 
           <div className='mt-14'>
-            <Button
-              variant='contained'
-              className='rounded-2xl px-8 py-4 text-title text-white'
-              href={EXPLORE_DATA_PAGE_LINK}
-            >
+            <HetBigCTA href={EXPLORE_DATA_PAGE_LINK}>
               Explore the data
-            </Button>
+            </HetBigCTA>
           </div>
         </article>
