+++ frontend/src/styles/HetComponents/HetHorizontalRuleHeader.tsx
